+++ examples/plugins/src/lighthouse/src/lighthouse.plugin.unit.test.ts
@@ -1,24 +1,16 @@
 import { describe, expect, it } from 'vitest';
-import { LIGHTHOUSE_URL } from '../mock/constants';
 import { LIGHTHOUSE_OUTPUT_FILE_DEFAULT } from './constants';
 import { runnerConfig } from './lighthouse.plugin';
-import type { LighthouseCliOptions } from './types';
 
 describe('lighthouse-runnerConfig', () => {
-  const baseOptions: LighthouseCliOptions = {
-    url: LIGHTHOUSE_URL,
-  };
-  const lcpAuditOutputBase = {
-    displayValue: expect.stringContaining('sec'),
-    score: 1,
-    slug: 'largest-contentful-paint',
-    value: 0,
-  };
-
   it('should execute if url is given', () => {
-    expect(runnerConfig(baseOptions)).toEqual(
+    expect(
+      runnerConfig({
+        url: 'http://localhost:8080',
+      }),
+    ).toEqual(
       expect.objectContaining({
-        args: expect.arrayContaining(['lighthouse', LIGHTHOUSE_URL]),
+        args: expect.arrayContaining(['lighthouse', 'http://localhost:8080']),
         command: 'npx',
         outputFile: LIGHTHOUSE_OUTPUT_FILE_DEFAULT,
       }),
@@ -26,7 +18,11 @@ describe('lighthouse-runnerConfig', () => {
   });
 
   it('should execute with output "json" and output-path "lighthouse-report.json" by default', () => {
-    expect(runnerConfig(baseOptions)).toEqual(
+    expect(
+      runnerConfig({
+        url: 'http://localhost:8080',
+      }),
+    ).toEqual(
       expect.objectContaining({
         args: expect.arrayContaining([
           '--output="json"',
@@ -42,13 +38,13 @@ describe('lighthouse-runnerConfig', () => {
   it('should run only audits included in given onlyAudits', () => {
     expect(
       runnerConfig({
-        ...baseOptions,
-        onlyAudits: [lcpAuditOutputBase.slug],
+        url: 'http://localhost:8080',
+        onlyAudits: ['largest-contentful-paint'],
       }),
     ).toEqual(
       expect.objectContaining({
         args: expect.arrayContaining([
-          `--onlyAudits="${lcpAuditOutputBase.slug}"`,
+          '--onlyAudits="largest-contentful-paint"',
         ]),
       }),
     );
@@ -56,7 +52,7 @@ describe('lighthouse-runnerConfig', () => {
 
   it('should parse options for headless by default to false', () => {
     const args = runnerConfig({
-      url: LIGHTHOUSE_URL,
+      url: 'http://localhost:8080',
     });
     expect(args).toEqual(
       expect.not.arrayContaining(['--chrome-flags="--headless=new"']),
@@ -66,12 +62,12 @@ describe('lighthouse-runnerConfig', () => {
   it('should run headless "new" if set to true', () => {
     expect(
       runnerConfig({
-        ...baseOptions,
+        url: 'http://localhost:8080',
         headless: 'new',
       }),
     ).toEqual(
       expect.objectContaining({
-        args: expect.arrayContaining([`--chrome-flags="--headless=new"`]),
+        args: expect.arrayContaining(['--chrome-flags="--headless=new"']),
       }),
     );
   });