+++ global-setup.e2e.ts
@@ -1,9 +1,6 @@
 import { execSync } from 'child_process';
-import {
-  setup as globalSetup,
-  teardown as globalTeardown,
-} from './global-setup';
-import { setupTestFolder } from './testing/test-setup/src';
+import { setup as globalSetup } from './global-setup';
+import { setupTestFolder, teardownTestFolder } from './testing/test-setup/src';
 import startLocalRegistry from './tools/scripts/start-local-registry';
 import stopLocalRegistry from './tools/scripts/stop-local-registry';
 
@@ -21,5 +18,6 @@ export async function teardown() {
   execSync('npm uninstall @code-pushup/cli');
   execSync('npm uninstall @code-pushup/eslint-plugin');
   execSync('npm uninstall @code-pushup/coverage-plugin');
-  await globalTeardown();
+  await teardownTestFolder('tmp/e2e');
+  await teardownTestFolder('tmp/local-registry');
 }