+++ packages/core/src/lib/implementation/collect.integration.test.ts
@@ -1,5 +1,6 @@
 import { vol } from 'memfs';
 import { describe, expect, it } from 'vitest';
+import { commitSchema } from '@code-pushup/models';
 import { MEMFS_VOLUME, MINIMAL_CONFIG_MOCK } from '@code-pushup/test-utils';
 import { collect } from './collect';
 
@@ -7,6 +8,7 @@ describe('collect', () => {
   it('should execute with valid options', async () => {
     vol.fromJSON({}, MEMFS_VOLUME);
     const report = await collect({
+      categories: [],
       ...MINIMAL_CONFIG_MOCK,
       verbose: true,
       progress: false,
@@ -27,5 +29,7 @@ describe('collect', () => {
         },
       }),
     );
+
+    expect(() => commitSchema.parse(report.commit)).not.toThrow();
   });
 });