+++ packages/plugin-coverage/src/lib/runner/index.ts
@@ -10,7 +10,7 @@ import {
 } from '@code-pushup/utils';
 import { FinalCoveragePluginConfig } from '../config';
 import { applyMaxScoreAboveThreshold } from '../utils';
-import { PLUGIN_CONFIG_PATH, RUNNER_OUTPUT_PATH, WORKDIR } from './constants';
+import { PLUGIN_CONFIG_PATH, RUNNER_OUTPUT_PATH } from './constants';
 import { lcovResultsToAuditOutputs } from './lcov/lcov-runner';
 
 export async function executeRunner(): Promise<void> {
@@ -20,14 +20,13 @@ export async function executeRunner(): Promise<void> {
   // Run coverage tool if provided
   if (coverageToolCommand != null) {
     const { command, args } = coverageToolCommand;
-
     try {
       await executeProcess({ command, args });
     } catch (error) {
       if (error instanceof ProcessError) {
-        console.info(chalk.bold('stdout from failed process:'));
-        console.info(error.stdout);
-        console.error(chalk.bold('stderr from failed process:'));
+        console.error(chalk.bold('stdout from failed coverage tool process:'));
+        console.error(error.stdout);
+        console.error(chalk.bold('stderr from failed coverage tool process:'));
         console.error(error.stderr);
       }
 
@@ -37,7 +36,7 @@ export async function executeRunner(): Promise<void> {
     }
   }
 
-  // Caculate coverage from LCOV results
+  // Calculate coverage from LCOV results
   const auditOutputs = await lcovResultsToAuditOutputs(reports, coverageTypes);
 
   await ensureDirectoryExists(dirname(RUNNER_OUTPUT_PATH));
@@ -49,7 +48,7 @@ export async function createRunnerConfig(
   config: FinalCoveragePluginConfig,
 ): Promise<RunnerConfig> {
   // Create JSON config for executeRunner
-  await ensureDirectoryExists(WORKDIR);
+  await ensureDirectoryExists(dirname(PLUGIN_CONFIG_PATH));
   await writeFile(PLUGIN_CONFIG_PATH, JSON.stringify(config));
 
   const threshold = config.perfectScoreThreshold;