+++ packages/plugin-eslint/src/lib/runner.integration.test.ts
@@ -1,15 +1,16 @@
 import { ESLint } from 'eslint';
-import { mkdir, rm, writeFile } from 'node:fs/promises';
+import { rm, writeFile } from 'node:fs/promises';
 import os from 'node:os';
 import { dirname, join } from 'node:path';
 import { fileURLToPath } from 'node:url';
-import { SpyInstance, describe, expect, it } from 'vitest';
+import { MockInstance, describe, expect, it } from 'vitest';
 import type { AuditOutput, AuditOutputs, Issue } from '@code-pushup/models';
 import { osAgnosticAuditOutputs } from '@code-pushup/test-utils';
-import { readJsonFile } from '@code-pushup/utils';
+import { ensureDirectoryExists, readJsonFile } from '@code-pushup/utils';
 import { listAuditsAndGroups } from './meta';
 import {
   ESLINTRC_PATH,
+  PLUGIN_CONFIG_PATH,
   RUNNER_OUTPUT_PATH,
   createRunnerConfig,
   executeRunner,
@@ -17,20 +18,14 @@ import {
 import { setupESLint } from './setup';
 
 describe('executeRunner', () => {
-  let cwdSpy: SpyInstance;
-  let platformSpy: SpyInstance;
+  let cwdSpy: MockInstance<[], string>;
+  let platformSpy: MockInstance<[], NodeJS.Platform>;
 
-  const createArgv = async (eslintrc: string) => {
+  const createPluginConfig = async (eslintrc: string) => {
     const patterns = ['src/**/*.js', 'src/**/*.jsx'];
     const eslint = setupESLint(eslintrc);
     const { audits } = await listAuditsAndGroups(eslint, patterns);
-    const runnerConfig = createRunnerConfig(
-      'bin.js',
-      audits,
-      eslintrc,
-      patterns,
-    );
-    return [runnerConfig.command, ...(runnerConfig.args ?? [])];
+    await createRunnerConfig('bin.js', audits, eslintrc, patterns);
   };
 
   const appDir = join(
@@ -50,7 +45,7 @@ describe('executeRunner', () => {
     const config: ESLint.ConfigData = {
       extends: '@code-pushup',
     };
-    await mkdir(dirname(ESLINTRC_PATH), { recursive: true });
+    await ensureDirectoryExists(dirname(ESLINTRC_PATH));
     await writeFile(ESLINTRC_PATH, JSON.stringify(config));
   });
 
@@ -59,41 +54,40 @@ describe('executeRunner', () => {
     platformSpy.mockRestore();
 
     await rm(ESLINTRC_PATH, { force: true });
+    await rm(PLUGIN_CONFIG_PATH, { force: true });
   });
 
   it('should execute ESLint and create audit results for React application', async () => {
-    const argv = await createArgv('.eslintrc.js');
-
-    await executeRunner(argv);
+    await createPluginConfig('.eslintrc.js');
+    await executeRunner();
 
     const json = await readJsonFile<AuditOutputs>(RUNNER_OUTPUT_PATH);
     expect(osAgnosticAuditOutputs(json)).toMatchSnapshot();
   });
 
   it('should execute runner with inline config using @code-pushup/eslint-config', async () => {
-    const argv = await createArgv(ESLINTRC_PATH);
-
-    await executeRunner(argv);
+    await createPluginConfig(ESLINTRC_PATH);
+    await executeRunner();
 
     const json = await readJsonFile<AuditOutput[]>(RUNNER_OUTPUT_PATH);
     // expect warnings from unicorn/filename-case rule from default config
     expect(json).toContainEqual(
-      expect.objectContaining({
+      expect.objectContaining<Partial<AuditOutput>>({
         slug: 'unicorn-filename-case',
         displayValue: '5 warnings',
         details: {
-          issues: expect.arrayContaining([
+          issues: expect.arrayContaining<Issue>([
             {
               severity: 'warning',
               message:
                 'Filename is not in kebab case. Rename it to `use-todos.js`.',
-              source: expect.objectContaining({
+              source: expect.objectContaining<Issue['source']>({
                 file: join(appDir, 'src', 'hooks', 'useTodos.js'),
-              } satisfies Partial<Issue['source']>),
-            } satisfies Issue,
+              }),
+            },
           ]),
         },
-      } satisfies Partial<AuditOutput>),
+      }),
     );
   }, 7000);
 });