+++ packages/utils/perf/score-report/optimized3.ts
@@ -1,13 +1,14 @@
 /* eslint-disable no-param-reassign, functional/immutable-data */
 // Note: The mutability issues are resolved in production code
-import { CategoryRef, GroupRef, Report } from '@code-pushup/models';
-import { ScoredReport } from '../../src';
 import {
-  EnrichedAuditReport,
-  EnrichedScoredGroup,
-  GroupRefInvalidError,
-  ScoredCategoryConfig,
-} from '../../src/lib/reports/scoring';
+  AuditReport,
+  CategoryRef,
+  GroupRef,
+  Report,
+} from '@code-pushup/models';
+import { ScoredReport } from '../../src';
+import { GroupRefInvalidError } from '../../src/lib/reports/scoring';
+import { ScoredCategoryConfig, ScoredGroup } from '../../src/lib/reports/types';
 
 export function calculateScore<T extends { weight: number }>(
   refs: T[],
@@ -44,18 +45,14 @@ export function deepClone<T>(obj: T): T {
 
 export function scoreReportOptimized3(report: Report): ScoredReport {
   const scoredReport = deepClone(report) as ScoredReport;
-  const allScoredAuditsAndGroups = new Map<
-    string,
-    EnrichedAuditReport | EnrichedScoredGroup
-  >();
+  const allScoredAuditsAndGroups = new Map<string, AuditReport | ScoredGroup>();
 
-  scoredReport.plugins?.forEach(plugin => {
+  scoredReport.plugins.forEach(plugin => {
     const { audits, slug } = plugin;
-    const groups = plugin.groups || [];
+    const groups = plugin.groups ?? [];
 
     audits.forEach(audit => {
       const key = `${slug}-${audit.slug}-audit`;
-      audit.plugin = slug;
       allScoredAuditsAndGroups.set(key, audit);
     });
 
@@ -72,7 +69,6 @@ export function scoreReportOptimized3(report: Report): ScoredReport {
     groups.forEach(group => {
       const key = `${slug}-${group.slug}-group`;
       group.score = calculateScore(group.refs, groupScoreFn);
-      group.plugin = slug;
       allScoredAuditsAndGroups.set(key, group);
     });
     plugin.groups = groups;