+++ testing/test-utils/src/lib/utils/dynamic-mocks/eslint-plugin.mock.ts
@@ -2,11 +2,27 @@ import { join } from 'node:path';
 import type {
   Audit,
   CategoryRef,
+  Group,
   PluginConfig,
   PluginReport,
 } from '@code-pushup/models';
 import { ESLINT_AUDITS_MAP } from './eslint-audits.mock';
 import { echoRunnerConfigMock } from './runner-config.mock';
+
+const ESLINT_PLUGIN_GROUP_MAX_LINES: Group = {
+  slug: 'max-line-limitation',
+  title: 'Maximum lines limitation',
+  refs: [
+    {
+      slug: ESLINT_AUDITS_MAP['max-lines-per-function'].slug,
+      weight: 1,
+    },
+    {
+      slug: ESLINT_AUDITS_MAP['max-lines'].slug,
+      weight: 1,
+    },
+  ],
+};
 
 const eslintMeta = {
   slug: 'eslint',
@@ -43,6 +59,7 @@ export function eslintPluginReportMock(): PluginReport {
     date: '2023-10-18T07:49:45.531Z',
     duration: 368,
     audits: Object.values(ESLINT_AUDITS_MAP),
+    groups: [ESLINT_PLUGIN_GROUP_MAX_LINES],
   };
 }